requirements.txt
ollama 
langdetect  
sentence-transformers  
chromadb  
flask  
python-dotenv 
